requirements.txt
Django==5.2.10
mysqlclient==2.2.8
djangorestframework==3.16.1
django-cors-headers==4.6.0
requests==2.32.3
